set/getId is in OperationalLimits
