+++ iidm/iidm-api/src/main/java/com/powsybl/iidm/network/LoadingLimitsAdder.java
@@ -10,9 +10,7 @@ package com.powsybl.iidm.network;
  *
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public interface LoadingLimitsAdder<L extends LoadingLimits, A extends LoadingLimitsAdder<L, A>> extends OperationalLimitsAdder<L> {
-
-    A setId(String id);
+public interface LoadingLimitsAdder<L extends LoadingLimits, A extends LoadingLimitsAdder<L, A>> extends OperationalLimitsAdder<L, A> {
 
     interface TemporaryLimitAdder<A> {
